requirements.txt
country_list==1.0.0
ecdsa==0.18.0
gmpy2==2.1.2
pydantic==1.9.1
SQLAlchemy==1.4.39
pyyaml==6.0